style(Manage) : service 구현 분리 ManageService 구현, 인터페이스 분리
